changing floats to int for damage

+++ Issues The Game/Assets/Scripts/EnemyHealth.cs
@@ -4,8 +4,8 @@ using UnityEngine;
 
 public class EnemyHealth : MonoBehaviour
 {
-    public float maxHealth = 100;
-    private float health;
+    public int maxHealth = 100;
+    private int health;
     // Start is called before the first frame update
     void Start()
     {
@@ -16,7 +16,7 @@ public class EnemyHealth : MonoBehaviour
     {
 
     }
-    public void TakeDamage(float damage)
+    public void TakeDamage(int damage)
     {
         health -= damage;
         if (health <= 0)
